+++ src/logrotate.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 from charms.operator_libs_linux.v1 import systemd
 from github_runner_manager.metrics.events import METRICS_LOG_PATH
-from github_runner_manager.reactive.runner_manager import REACTIVE_RUNNER_LOG_DIR
+from github_runner_manager.reactive.process_manager import REACTIVE_RUNNER_LOG_DIR
 from pydantic import BaseModel
 
 from errors import LogrotateSetupError