+++ docs/changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+### 2024-09-30
+
+- Improve reconciliation for reactive runners
+
+### 2024-09-27
+
+- Add job label validation when consuming a job from the message queue.
+
+### 2024-09-24
+
+- Add support for spawning a runner reactively
+- Fix a bug where busy runners are killed instead of only idle runners
+
 ### 2024-09-18
 
 - Changed code to be able to spawn a runner in reactive mode.